WIP: Rework entire-line implementation and add test coverage for combination with literal matching
For a single line match, we also need an end-of-line assertion for --line-regexp. For subsequent lines, the assertion usually is "any non-newline characters", to ensure that the pattern indeed matches in the last (current) line, and not accidentally in a previous line. With --line-regexp, that needs to go, to anchor at the end. Any lines in the middle don't need the $toEndOfLineAsssertion (it didn't do any harm, though).