echotrace: Improve eval handling
The simple dropping of the "eval" itself is not useful, as the following argument are re-quoted and therefore cannot be copied and executed as-is. Instead, drop the special eval handling by default (so that the result, though perhaps visually unappealing, can be directly copied and executed), and offer a --eval mode that prints the evaluated result (i.e. a long command-line or a scriptlet across multiple lines) _if_ the following args are indeed an eval argument.