userchoice: ENH: Also consider charLimit, handle newlines in nchars reading
When echoing the typed input, the charLimit has to be considered when deciding whether there's wrapping in the terminal (and therefore the multi-line clearing algorithm has to be used). As the choice may be shorter than the maximum possible length (when a -d delimiter is encountered early), the actual anticipated line wrapping may not have happened yet, and the clearing would affect one line too many. To avoid that, force wrapping by appending a padding of charLimit spaces and backtracking via ^H to the original position, and accounting for a too short input via extra padding before clearing.
Newlines in echoed input pose another challenge, as these increase the number of lines to be cleared. These get recorded (across retries due to invalid input) in newlinesInChoice, and then cause the cursor to go up more lines to counterbalance the newlines. We need to see all entered newlines for this to work, so userquery is reconfigured (via USERQUERY_EOF) to append a "$" instead of the newline (that gets chomped by the shell together with any trailing whitespace).