userquery: Improve nchars reading of newlines
Clearing IFS avoids that leading newlines are missing from the choice, and so the whole read text is returned.
A client can influence the output format of $USERQUERY_EOF; e.g. by setting it to '$' the chomping off of trailing whitespace can be avoided (and the client has to cut off the trailing '$' itself).