userchoiceFromList: ENH: Allow case-insensitive accelerators via -i|--ignore-case
For very simple choices (like [Y]es/[N]o), uppercase accelerators both make the choices easier to visually parse and allow for more flexibility in input (e.g. if Caps Lock is accidentally pressed, or a previous choice required uppercase and the finger is slow to move off the Shift key).
Accelerators are normalized to lowercase then, both variants are added as valid inputs, and the chosen accelerator again is lowercased before looking it up.