ENH: Allow conversion to POSIX syntax via --posix
This is useful when a Bash script prepares a command-line to be executed by a different program (that uses the default (POSIX) shell) - e.g. AWK's "commmand" | getline or sed's "e COMMAND". The $'...' sequences generated by Bash's %q cannot be interpreted:

$ bash -c "echo this is \$USER \$'foo\\tbar'"
this is inkarkat foo    bar
$ sh -c "echo this is \$USER \$'foo\\tbar'"
this is inkarkat $foo   bar