Reimplementation: Avoid special anchor handling by using split() instead of patsplit()
By turning everything around and using the pattern as the separator, not the matched content, the problem with anchors can be avoided altogether. (So much for spending all morning to get the escaping right :-(
Matches now split on its pattern and use the separators = non-matches for extracting the matched text.
Skips now split on its pattern and use the matches = everything that did not match.