Do not print separator for single omitted line
Instead of omitting a single line and replacing it with a line that mentions that the line was omitted, print the NUM+1 directly. It's less distracting and one "bonus" line is also more useful for the user. And it avoids any plural handling in HEAD-TAIL-SEPARATOR, as the amount of omitted lines will now always be 2 or higher.