Implement --{get,remove}-both
There's an added complication here with the potential overlap of prefix and suffix if lines are completely identical or have identical prefixes and suffixes. "foobar" "foobar" have the same prefix "foobar" and no suffix at all; "foobarfoobar" "foobar-foobar" have the same prefix "foobar" and suffix "foobar". This matters when removing them!
We solve this by storing the minimum start position of the suffix; if that one overlaps with the prefix, the corresponding text has to be dropped from the suffix.