uniqueStable: ENH: Allow output of last vs. first occurrence, and even both
The last occurrence can be obtained via $ tac | uniqueStable | tac, but that's a lot of pipelining (and keeping the whole contents in memory); we can do this more efficiently within the AWK program (that so far is just a one-liner for the uniqueness; most of the code deals with the input processing, and all of that can be leveraged).
Likewise, both first and last occurrences can be obtained via two separate passes, but apart from the efficiency loss, the original ordering then gets lost as well.