headtail: Consistency: Also handle --lines 0
This corner case so far caused a division by zero error, which is not nice.
Extract an algorithm (which will also be useful for inverted output with negative counts).
We can use head (or tail) as the only functionality is the file header; the contents are all suppressed by the line count 0.