trace: BUG: Regression: traps don't see script variables any longer
I noticed when my temp file cleanup didn't remove any of the files (stored in an array). The sourceIt() function opens a scope that a typeset -a is now bound to, but the tracee's trap runs in the global scope, where the function scope has been destroyed already.

,----[ tracee ]----
| typeset -a tempFiles=(foo bar baz)
| trap 'dump-args -a tempFiles -- "${tempFiles[@]}"' EXIT
`----
With a plain scalar variable, this doesn't happen, because I usually don't "typeset" them (and "local" would fail in the normal execution). Only my style of explicitly declaring array variables causes the problem.

Inline the sourceIt() function; it's been extracted purely for readability and is only used once, anyway.