ENH: Use URL-like escaping in combination with hash
We want as much recognizability of the string as possible, so don't simply use a hash if the text is too long.
Instead, combine escaped parts with hash, and adapt the hash length if the maximum length (now configurable via -l|--max-length) is aggressive.

System-info: Ubuntu 20.04.6 LTS, x86_64
Platform-info: bash 5.0.17(1)-release, Bats 1.10.0