ENH: Remove the captured input if the full contents have been printed
No need to leave behind a (potentially large and sensitive) temp file if its name has never been mentioned (in a cutline that wasn't printed).
Note: For simplicity, the AWK script only indicates whether _any_ truncation has happened, not if it happened on stdin in particular. This means that if FILE(s) other than stdin are passed as well (already unlikely), and those are truncated but stdin itself isn't, the temp file is kept even though it could be removed. We'll accept that small fuzziness to keep the implementation more understandable.

System-info: Ubuntu 20.04.6 LTS, x86_64
Platform-info: bash 5.0.17(1)-release, Bats 1.10.0