ENH: UI: Robustness: Wrap and truncate long notification text by default
The command execution can potentially create an awful lot of text, which most sinks (GUI dialog and desktop notification) cannot properly display, anyway, and would simply overload the user.
As the notification text is passed as an command-line argument, it's affected by shell limits, too, and we need to avoid the command aborting because of that.
We cannot control what format the output is; it could be many lines, but also a few overly long unbroken paragraphs. Therefore, do both word-wrapping (via fmt) and text truncation (via headtail). The latter command not only tries to remove the least important middle section, it now conveniently also captures the full output and adds a reference to the temp file in its cutline, so the user is able to copy-and-paste / find that file if they are actually interested in the full output. (For example, maybe typically the command just prints a short message, but now is has somehow errored out and dumped a long stacktrace.)

System-info: Ubuntu 20.04.6 LTS, x86_64
Platform-info: bash 5.0.17(1)-release, Bats 1.10.0